Add support for slash commands
The method setup() now returns a CommandTree object `tree`, which is
awaited in `on_ready()`. Use `app_commands` instead of `commands` to add
the commands to the tree in general.py instead of directly adding them
to the bot.

These changes add the slash commands to the list of slash commands on
discord, but now the functions have to be rewritten in incorporate the
new incoming data from the slash commands. They need to be rewritten
anyways.